requirements/base.txt
# Django
Django==3.2.2
django-environ==0.4.5

# Pandas
pandas==1.2.4

# Static files / Images
Pillow==8.2.0
django-versatileimagefield==2.0
whitenoise==5.2.0

# REST Framework
django-cors-headers==3.5.0
django-filter==2.4.0
django-guardian==2.3.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0

# Database
psycopg2-binary==2.8.6

# Webserver
gunicorn==20.1.0

# Misc
asgiref==3.3.4
install==1.3.4
Markdown==3.3.4
Pygments==2.9.0
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1


